third_party/blink/web_tests/flag-specific/highdpi/virtual/container-queries/wpt_internal/css/css-conditional/container-queries/block-size-containment-expected.txt

This is a testharness.js-based test.
FAIL block-size containment only assert_equals: expected 50 but got 20
Harness: the test ran to completion.
